+++ spec/classes/init_spec.rb
@@ -2,18 +2,26 @@ require 'spec_helper'
 
 describe 'dhcp' do
   context 'supported operating systems' do
-    ['Debian', 'RedHat', 'FreeBSD'].each do |osfamily|
-      conf_path = (osfamily == 'FreeBSD') ? '/usr/local/etc' : '/etc/dhcp'
-      describe "dhcp class without any parameters on #{osfamily}" do
+    on_supported_os.each do |os, facts|
+      conf_path = case os
+                  when /^FreeBSD/i
+                    '/usr/local/etc'
+                  when /^Archlinux/i
+                    '/etc'
+                  else
+                    '/etc/dhcp'
+                  end
+      describe "dhcp class without any parameters on #{os}" do
         let(:params) do {
           :interfaces => ['eth0'],
         } end
 
-        let(:facts) do {
-          :concat_basedir => '/doesnotexist',
-          :domain         => 'example.org',
-          :osfamily       => osfamily,
-        } end
+        let(:facts) do
+          facts.merge({
+            :concat_basedir => '/doesnotexist',
+            :domain         => 'example.org',
+          })
+        end
 
         it { should compile.with_all_deps }
 
@@ -37,7 +45,7 @@ describe 'dhcp' do
         }
       end
 
-      describe "dhcp class parameters on #{osfamily}" do
+      describe "dhcp class parameters on #{os}" do
         let(:params) do {
           :interfaces   => ['eth0'],
           :dnsupdatekey => 'mydnsupdatekey',
@@ -45,7 +53,14 @@ describe 'dhcp' do
           :omapi_name   => 'mykeyname',
           :omapi_key    => 'myomapikey',
           :pxeserver    => '10.0.0.5',
+          :mtu          => 9000,
           :pxefilename  => 'mypxefilename',
+          :bootfiles    => {
+            '00:00'       => 'pxelinux.0',
+            '00:06'       => 'shim.efi',
+            '00:07'       => 'shim.efi',
+            '00:09'       => 'shim.efi',
+          },
           :option_static_route => true,
           :options      => ['provision-url code 224 = text', 'provision-type code 225 = text'],
           :authoritative => true,
@@ -54,11 +69,12 @@ describe 'dhcp' do
           :includes => ['myinclude1', 'myinclude2'],
         } end
 
-        let(:facts) do {
-          :concat_basedir => '/doesnotexist',
-          :domain         => 'example.org',
-          :osfamily       => osfamily,
-        } end
+        let(:facts) do
+          facts.merge({
+            :concat_basedir => '/doesnotexist',
+            :domain         => 'example.org',
+          })
+        end
 
         it { should compile.with_all_deps }
 
@@ -94,14 +110,64 @@ describe 'dhcp' do
             'option pxegrub code 150 = text ;',
             'option rfc3442-classless-static-routes code 121 = array of integer 8;',
             'option ms-classless-static-routes code 249 = array of integer 8;',
+            'option interface-mtu 9000;',
             'option provision-url code 224 = text;',
             'option provision-type code 225 = text;',
             'next-server 10.0.0.5;',
-            'filename "mypxefilename";',
+            'option architecture code 93 = unsigned integer 16 ;',
+            'if option architecture = 00:00 {',
+            '  filename "pxelinux.0";',
+            '} elsif option architecture = 00:06 {',
+            '  filename "shim.efi";',
+            '} elsif option architecture = 00:07 {',
+            '  filename "shim.efi";',
+            '} elsif option architecture = 00:09 {',
+            '  filename "shim.efi";',
+            '} else {',
+            '  filename "mypxefilename";',
+            '}',
             'log-facility local7;',
             "include \"#{conf_path}/dhcpd.hosts\";",
             'include "myinclude1";',
             'include "myinclude2";',
+          ])
+        }
+      end
+
+      describe "ddns-updates without key" do
+        let(:params) do {
+          :interfaces => ['eth0'],
+          :ddns_updates => true,
+        } end
+
+        let(:facts) do
+          facts.merge({:domain => 'example.org'})
+        end
+
+        it { should compile.with_all_deps }
+
+        it {
+          verify_concat_fragment_exact_contents(catalogue, 'dhcp.conf+01_main.dhcp', [
+            'omapi-port 7911;',
+            'default-lease-time 43200;',
+            'max-lease-time 86400;',
+            'ddns-updates on;',
+            'ddns-update-style interim;',
+            'update-static-leases on;',
+            'use-host-decl-names on;',
+            'zone example.org. {',
+            '  primary 8.8.8.8;',
+            '}',
+            'option domain-name "example.org";',
+            'option domain-name-servers 8.8.8.8, 8.8.4.4;',
+            "option ntp-servers none;",
+            'allow booting;',
+            'allow bootp;',
+            'option fqdn.no-client-update    on;  # set the "O" and "S" flag bits',
+            'option fqdn.rcode2            255;',
+            'option pxegrub code 150 = text ;',
+            'log-facility local7;',
+            "include \"#{conf_path}/dhcpd.hosts\";",
           ])
         }
       end