+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 3.1.0
+* Add `$ddns_updates` parameter to allow insecure DDNS updates.
+* Add `$pxefilename` parameter to `dhcp::pool`.
+* Add `$raw_prepend` and `$raw_append` parameters to `dhcp::pool`.
+* Fix classless static routes.
+
+## 3.0.0
+* Add bootfiles parameter with hash of client architectures to boot
+  loaders used for "filename", defaulting to "pxelinux.0" (#14920)
+* Add mtu parameter to main class and to dhcp::pool
+* Add Arch Linux support
+* Drop support for Ruby 1.8.7
+* Many improvements to tests
+
 ## 2.3.2
 * Fix metadata to show Puppet 4 compatibility
 * Remove hashes around pool names for Webmin compatibility